[cnf] expose ~toplevel flag in make_form so t. make_term can set it correctly
This is a more global fix for issue
https://github.com/OCamlPro/alt-ergo/issues/290. Ie, that works even for
cases where functions declarations return values of type 'a that becomes
bool once type-checking is done. The assert failure in the issue happens
because function f(X) = (forall Y...) is transformed to:
forall X. f(X) = (forall Y...). But when building hash-consed expressions
of the equality, the call to make_from from make_term to process the
internal (forall Y ...), the 'toplevel' flag was implicitely true. So, all
type variables appearing in it are bound in this quantification, which is
not correct