[bugfix] defining terms of functions/predicates extracted from typed AST
Defining terms of functions and predicates was extracted from hashconsed
expressions. But this is not optimal, and leads to a bug in case of
recursive functions / predicates.

A simpler workaround is to extract defining terms from typed AST, which
is more syntactic: no simplifications are made, and we know exactly where
to find these defining terms