Add support for --parse-only and --type-only to dolmen frontend
This patch adds support for the `--parse-only` and `--type-only` options
on the dolmen frontend. If these options are passed, the Dolmen_loop
pipeline is stopped early using the `@>|>` operator.

At the same time, the PR cleans up the dolmen pipeline usage by removing
calls to the `Header` module that is not meant to be used by Alt-Ergo
according to @Gbury.