Restore floating-point prelude support
Adding int parameters to `Sy.Float` in #596 involuntarily broke the
floating-point prelude. This is because when we build the expressions,
the parameters of the floating-point constructor may be variables, which
then get instantiated by actual constants in a later phase.

This restores the old behavior, where we store the
precision/exponent/mode arguments as expressions, and convert them into
constants lazily post-instantiation.