Better pre-commit hook
The current pre-commit hook in rsc/extra/pre-commit--git-hook is fairly
frustrating to use for multiple reasons:

 - It always checks *all* the files in the repository, making it slow;
 - It prints messages even when everything is fine (if I remember
   correctly, not been using it for a long time)
 - It always check against the working copy, not the index, meaning that
   trying to commit twice (without adding the changes in between) always
   work.

This patch provides an alternative implementation of a pre-commit hook
that:

 - Only checks the files that will be modified by the tentative commit
 - Is silent when all checks pass
 - Checks against the index, ensuring that it will never let you commit
   ill-indented files (well, unless you use `--no-verify` of course).

For indentation errors, if there are any, the script re-indents the
corresponding file(s) using `ocp-indent`, and tells you about the
changes. You must add (or not) the corresponding changes manually. This
is required to be compatible with workflows based on partial commits
(using `git add -p`).

For whitespace errors, the script simply checks them, and does not let
you commit if there are any, because that is provided for free by `git`.
Whitespace errors are easy to avoid with proper editor configuration,
and as an alternative a `.editorconfig` file is provided to help with
editor configuration.