Remove pinning of dolmen to git version
This was done because we needed support for mutually recursive
definitions, but that is present in dolmen 0.8.1 which is the currently
released version.

On the other hand, the dev version of Dolmen has introduced
backwards-incompatible API changes (see #629), so we can't use it
anymore.

This removes the "Temporary fix" in the GitHub Actions and uses the
publicly released dolmen version everywhere.