Incorporate the FPA prelude as a builtin
This patch adds support for "builtin" theory preludes, i.e. preludes
that are embedded in the Alt-Ergo binary (actually, library) rather than
being bundled as separate files.  The FPA prelude is moved from an
"old-style" prelude to such a bundled prelude -- in fact, to three
separate preludes (RIA, NRA, and FPA) to better reflect the fact that
some of the axioms within are not specific to floating point arithmetic
at all.

This new style of preludes reflects the idea that the fact that
floating-point support is implemented using a prelude is just that: an
implementation detail, and is otherwise irrelevant. It could as well be
a builtin theory, and there should be no difference to the end user.

The `--prelude` option is preserved, both for backwards compatibility
but also because the concept of user-defined preludes is useful. The new
builtin preludes are controlled with the `--enable-preludes`,
`--disable-preludes`, and `--disable-default-preludes` option.

Note that, keeping in sync with the remark above, we may want to call
these new options `--enable-theories` / `--disable-theories` (although
we can't really have `--disable-default-theories`), but if we do so, we
should consolidate somehow to expose consistent options with the
existing `--disable-adts` and co (I think having a single option to
enable/disable the theories selectively is better than adding
`--disable-fpa`, `--disable-ria` and `--disable-nra`, but I could be
convinced otherwise).