Use dune-site instead of configure.ml
This patch removes the old configure.ml script and uses dune-site, which
are dune's built-in mechanism for loading data files. As part of the
transition, the preludes are now stored in `share/alt-ergo/preludes`
instead of `lib/alt-ergo/preludes`, which makes more sense given that
they are data files and not library files. The plugins are still in
`lib/alt-ergo/plugins`.

Besides removing a dependency on bespoke scripts, the use of dune-site
allows external plugins to register themselves to the alt-ergo preludes
directory. Not depending on the executable path also means that the
prelude and plugin loading work better when developing alt-ergo.