[frontend:legacy] Explicitely disable polymorphic enums
Polymorphic enums are not supported by the legacy typechecker (enums aka
sum types are always monomorphic), but they fail with a cryptic error
message (see #517).

The dolmen frontend deals with this properly, and we plan on deprecating
the legacy typechecker in the future, so it's probably not worth fixing
-- but at the very least we should have a more informative error
message.

Fixes #517