Represent bitvector variables using terms
The reasoning for this patch is explained
[here](https://github.com/OCamlPro/alt-ergo/pull/472#issuecomment-1498710402):

> The issue comes from the way Alt-Ergo manages `uninterpreted semantic values` in the Shostak module. By `uninterpreted semantic value`, I mean a semantic value whose the top symbol is uninterpreted. For instance:
>
>     * `x`
>
>     * `f(34)`
>
>     * `f(g(5)+5)`
>
>
> But `x+y` doesn't fall into this category since its top symbol is interpreted in the `Arith` theory (even if, it is not interpreted in the usual sense since `x` and `y` are variables).
>
> Now, you have two important functions defined in the Shostak module: `solve` and `leaves`. The `leaves` function returns the direct uninterpreted subvalues of a semantic value. For instance:
>
>     * `leaves(x) = []`
>
>     * `leaves(f(g(5)+5)) = [g(5)]`
>
>
> The function `solve` tries to solve a system of equations between semantic values and in case of success, it returns a sequence of substitutions. More precisely, these substitutions replace some leaves in the members of the equalities by semantic values. So the substitutions are of the form : ` r1 --> R1, r2 --> R2, ...` where `r1, r2, ...` are _uninterpreted_ semantic value.
>
> Unfortunately, the Bitvector theory produces wrong substitutions. The module produces substitutions whose the left member is interpreted. By miracle, these terms are not necessary to solve the equations, so if you break the invariant above, AE successes.
>
> Actually, it makes sense that Bitv produces such substitutions. The module has it own variable type. For the global theory, a variable produces in Bitv looks like an interpreted semantic value of Bitv and the Shostak module has no way to know that it´s actually a variable since the type of semantic value of Bitv is abstract.
>
> I had notice this strange design before. I am thinking now how to change it properly.

This patch associates with each bitvector variable a term, which is a
skolem constant representing the variable (note that we must use a
`Symbols.Name`, not a `Symbols.Var` here, even though we are
representing a "variable" for the bitvector theory, because
`Symbols.Var` is used *only* for the bound variables of quantifiers).
This skolem constant is the one that is used to communicate with the
Shostak theory: we advertise it as a leaf (previously, we were returning
bitvector-theory-level terms as leaves, which is incorrect, because
leaves *MUST* only be terms with an uninterpreted top symbol), and we
accept it as input of a substitution.

The uid of the term is used for comparisons instead of the old variable
counter. Although the uid is not identical (there are many terms
that are not bitvector variables), the uids for fresh variables are
guaranteed to be strictly increasing (a fresh variable can't, by
definition, be already present in the weak table), and so the existing
logic about variable ordering should not change.

Note that there may be some cases where we treat a bitvector variable as
an alien, and that could cause a slight change in behavior; however, I
don't believe this can actually happen because the term representatives
can only appear as leaves to a bitvector expression, and so they should
not be manipulated by the Shostak combiner without going through us.

Fixes #350
Fixes #664