Proper C-bitvector substitutions
When solving bitvector equalities, if there is a C-substitution that is
found, applying that substitution will always result in an assertion
failure, because we are iterating over a list recursively and have an
unconditional `assert false` in the empty case.

I believe that returning an empty list here is the correct thing, and
makes the function an actual substitution.

Note that in many cases we don't hit the assertion failure because if
there is no substitution at all, the `apply_subs` function is not called
in `equations_slice` (if I had to wager a guess, I would say that the
`assert false` in `apply_subs` came from trying to ensure that the
`subs` was not empty -- except it was not applied to the right
variable).

Fixes #649