Associate syntactic terms to bitvector variables
This is a new version of #679, which itself was a new version of #472.
A similar fix is also in #553 as bb554381a142600b15bde96adcc3a9c0813edd34

The crux of it is that the bitvector theory returns *interpreted* values
as leaves, where the rest of the solver requires leaves to be
*uninterpreted terms*.

The main difference of this patch with the other patches above is that
here we only create the fresh terms associated with variables *after*
solving, so we only create fresh terms that we do expose. On the other
hand, the solver may create many variables internally that gets split or
otherwise merged and whose associated terms would never get used.

Fixes #350
Fixes #664