Compatibility with new Dolmen constructs (#687)
Now that Gbury/dolmen#162 and Gbury/dolmen#166 have been merged the
build is broken. This patch removes the old In_interval constructor. It
was not working properly anyways. Support for the new Semantic_trigger
constructor that replaces it in dolmen will land in #681.

In addition, this patch is adapted to propagate attributes from the
typed statements. This will help determine theory and case-split
information in #662.