Make fpa_rounding_mode an ADT again (#672)
In #647 the fpa_rounding_mode was converted from an ADT to a builtin
type.  This broke the use of [is_theory_constant] on the
fpa_rounding_mode type, because it is no longer a theory constant (it
previously was a theory constant for the ADT theory, but being a builtin
it no longer has an associated theory).

This patch makes the fpa_rounding_mode type an ADT again.