Cleanup release process in preparation of 2.5.0 (#654)
* Remove "Changes" header from CHANGES.md

This causes issues with dune-release.

* Move parser loading to the CLI

This patch moves the `--add-parser` plugin loading mechanism from the
alt-ergo-parsers library to the CLI, where it is actually used. This is
cleaner architecturally (e.g. the plugin loading never worked on the JS
version), but also necessary to migrate to the dune-site mechanism.

* Move prelude resolution to the cli

This patch moves the prelude resolution and the preludesdir mechanism
from the parsers to the CLI. This way, we can give an error earlier if
the user provides an invalid prelude file, and we can share the prelude
resolution between the different frontends (currently the Dolmen
frontend is not doing prelude resolution).

* Move MyDynlink to bin/common

This patch moves the dynamic library loading MyDynlink from alt-ergo-lib
to the CLI. This means it can now live in the alt-ergo package, which is
required to use dune-site (otherwise we would have circular dependencies
between alt-ergo and alt-ergo-lib).

* Use dune-site instead of configure.ml

This patch removes the old configure.ml script and uses dune-site, which
are dune's built-in mechanism for loading data files. As part of the
transition, the preludes are now stored in `share/alt-ergo/preludes`
instead of `lib/alt-ergo/preludes`, which makes more sense given that
they are data files and not library files. The plugins are still in
`lib/alt-ergo/plugins`.

Besides removing a dependency on bespoke scripts, the use of dune-site
allows external plugins to register themselves to the alt-ergo preludes
directory. Not depending on the executable path also means that the
prelude and plugin loading work better when developing alt-ergo.

* Extract AB-Why3 to a separate plugin

This patch introduces a new opam package for the AB-Why3 plugin. This
plugin is separate from the rest of Alt-Ergo, and distributed under a
different license.

* Remove build date from version.ml

The build date is not available with dune-release. Besides, it is bad
for reproducibility to include the build date in the generated binaries.