Add support to (check-sat-assuming) in Dolmen frontend
This changes the way we parse the ```Solve`` constructor from Dolmen.

Currently, we accept the following forms:

 - ```Solve ([], [t])`, i.e. `goal g: t` in the native format
 - ```Solve ([t], [])`, i.e. `check_sat t` in the native format, and
   `(check-sat-assuming (t))` in the smtlib2 format
 - ```Solve ([], [])`, i.e. `(check-sat)` or `(check-sat-assuming ())`
   in the smtlib2 format

This patch adds support for the more general case of
`(check-sat-assuming)` by handling ```Solve (ts, [])`.

This is done by making the ```Check`` constructor take a list of term
(rather than a single term), and generalizing the `pp_query`
preprocessing helper in `D_cnf` to accept an arbitrary list of
hypotheses.

(Note that this means `(check-sat)` now has `Sat` as a `goal_sort`
rather than `Thm`, which is conceptually more correct but doesn't really
make a difference in practice)

Fixes #695