[models] Actually print values, not terms
This patch makes it so that the model values that we print are actually
values, not terms (i.e. they do not contain variables).  This is
achieved by aggressively substituting variable definitions when
printing.

The patch is a bit more involved than I'd have liked, because we can't
simply substitute semantic values into terms, so the guts of the model
printing had to be changed a little.

The code now works in two pass. In the first pass, we build a hash table
from symbols (constants and arrays) to their respective definitions
using the new `value_defn` mini-language, which we then use for
printing. The substitution is performed on-the-fly in the second pass,
which actually prints, using the hash table that was built in the first
pass.