Use --enable-theories instead of --enable-preludes
This patch removes the `--enable-preludes` and `--disable-preludes` flags
that leak implementation details and replaces them with
`--enable-theories` and `--disable-theories` flags that are more
user-friendly.

The flags `--no-ac` and `--disable-adts` are now deprecated aliases for
`--disable-theory ac` and `--disable-theory adt`, respectively.

The `--disable-default-preludes` flag has been removed, because it would
be confusing for users to have a `--disable-default-theories` flag that
only disables the prelude theories, ADT, and AC but not other theories
such as bitvectors. In the future, we should migrate all the theories to
the `--enable-theory` / `--disable-theory` flags and then we can
reintroduce the flag, which would behave similarly to `--restricted`
mode.

Fix #673