[refactor] Always represent classes as sets
Most representations of UF classes uses Expr.Set.t, but `class_of`
specifically converts the class to a list. There doesn't seem to be any
real reason for that, and it simply causes a bit of memory overhead by
creating the list instead of iterating over the expression set directly.

This patch is a first step towards abstracting the representation of UF
classes by using a common type for all their representations. It is not
entirely trivial to abstract the representation, because of cyclic
dependencies with `Explanation.Inconsistent`.