Run tests with dolmen by default
We plan on making dolmen the default frontend in the near future, so it
onyl makes sense that we actually start testing dolmen more.

This also simplifies slightly the handling of the 'dolmen' modifier,
since it only means that we must exclude the 'legacy' case now.

This also simplifies the test generation a bit by including the runtest
and runtest-quick targets on the CI, which means we don't need
duplicates for `tableaux` and `tableaux_ci`, etc anymore.