[bitv] Rewrite the canonizer (#732)
This is a full rewrite of the Canonizer module in the bitvector solver.
The new Canonizer (now called Canon) performs the same work as the
old one, but does so slightly differently -- in particular, we now build
terms lazily, which allows us to ignore parts of the tree for
performance.

The main motivation is actually not performance, but the ability to
integrate bitwise operators (bvnot, bvand, bvor) into the canonizer in a
more natural way. This is essentially the same architecture as that of
`make_aux` from #669, but lifted to use a simple custom type (`view`) to
avoid building new terms during canonization and benefit from match
exhaustiveness checks.