Expose the bv2nat and int2bv primitives (#743)
This patch exposes the bv2nat and int2bv primitives as symbols available
to the user. Even though these symbols are not part of the official
smtlib theories, they are useful for mixed bitvector/integer
specifications that are likely to occur in the context of program
verification, and they are also supported by other solvers such as Z3
and CVC5.