Restore record printing functionality (#793)
This got lost somewhere along the way as the conditions for model
generation changed.

Note that it is a bit weird that the accumulation of record accesses
happens in `models.ml` while the accumulation of array accesses happens
in `uf.ml`.  We should pick a single one for both (probably in `models`
but even better would be to expose the appropriate APIs so that each
theory can build its own model values), but for now this patch should
wrangle the `if`s properly and restore the existing functionality.