Ensure that internal AC symbols do not end up in models
The AC(X) algorithm uses special symbols as an indirection (I believe to
help enforce a terminating rewrite order) for AC symbols appearing
within other AC symbols. These internal symbols start with `@` which is
otherwise not a valid first character for a symbol name (but see #807).
They should not appear in models.