Use a separated counter for abstract value (#835)
* Use a separated counter for abstract value

Use a separated counter in `Models` to produce fresh names
for abstract values in models. This fix is temporary as we plan to
refactor the model generation in another PR.

A better design consists in producing fresh names while computing
the model.

* Increase the counter

* Promote tests

I also add another tests to check if the counter increases properly.