Use ZArith-based representations for numbers and bitvectors
Alt-Ergo uses strings to represent terms containing numbers and/or
bitvectors. In addition to a (slight) overhead, it also caused issues in
the past with different string representations of the same value being
considered distinct.

This patch uses Z.t (Q.t for reals) to represent integers, bitvectors
and reals instead of strings.

Fixes #668