[bitv] Add support for (bvnot) in the solver
This patch adds support for the negation (bvnot) operator in the
bitvector solver.

There are two components to this support:

 - First, the union-find data structure used by the solver is replaced
   with an actual union-find data structure implemented using Tarjan's
   efficient algorithm, rather than using an implementation with sets
   and maps. This is not only more efficient but also simplifies the
   implementation of the second part below.

 - Second, the new union-find data structure is augmented with a link
   between a class representative and the representative of its negated
   class. When merging two classes, their negated classes are merged as
   well (and, in particular, if a class gets forced to all ones or all
   zeroes, its negated classes gets forced to the other bit value). If a
   class is ever merged with its negated class, the problem is
   unsolvable.

Note that this also fixes a bug in the existing solver: previously, it
was possible to make bogus substitutions by asserting an equality
between a constant and a bit-vector term that depends on the constant,
for instance, `(= x (concat ((_ extract 0 0) x) ((_ extract 1 1) x)))`.
This would create the mappings `x -> x[0] @ x[1]` and `x[0] @ x[1] -> x`
in the `repr` table of the union-find, which is obviously incorrect and
confuses Alt-Ergo (see the `coherence` test).