Support optimize non-regression tests
We cannot properly run non-regression tests with only a subset of the solvers
defined in the main function of `gentest.ml`. I do a temporary ugly trick to
prevent running optimize tests with inappropriate solvers. This should be revert
as soon as Dolmen can parse a syntax for optimize constraints.