Split case_split handler into two parts
Refactoring of the function `case_split`. OptimAE introduces
a new kind of case_split: `optimized_split`. The previous code
manages all the case splits in a single function even if AE
processes optimized ones in a complete different way. This commit
divides the function `case_split` into two functions:
- `case_split` mostly as it was before OptimAE
- `optimizing_split` which try to optimize one case split by call.

We dispatch the call of `optimizing_split` in the `Relation` module
based on the type of the split to optimize.